content/content_v2/requirements.txt
anthropic
google-generativeai
python-dotenv
requests
Pillow
instaloader
pydantic
